refactor: centralize XML type helpers in XMLTypeData
