+++ core/src/main/kotlin/io/specmatic/core/pattern/XMLTypeData.kt
@@ -1,9 +1,16 @@
 package io.specmatic.core.pattern
 
 import io.specmatic.core.Resolver
+import io.specmatic.core.value.localName
+import io.specmatic.core.value.namespacePrefix
 import io.specmatic.core.value.StringValue
 import io.specmatic.core.value.XMLNode
 import io.specmatic.core.wsdl.parser.message.*
+
+internal const val XML_SCHEMA_INSTANCE_NAMESPACE = "http://www.w3.org/2001/XMLSchema-instance"
+internal const val XML_SCHEMA_NAMESPACE = "http://www.w3.org/2001/XMLSchema"
+
+internal data class WSDLTypeName(val namespace: String, val localName: String)
 
 data class XMLTypeData(
     val name: String = "",
@@ -105,5 +112,64 @@ data class XMLTypeData(
         return attributes[NILLABLE_ATTRIBUTE_NAME].let {
             it is ExactValuePattern && it.pattern.toStringLiteral().lowercase() == "true"
         }
+    }
+
+    internal fun xsiTypeName(): WSDLTypeName? {
+        val attribute = attributes.keys.firstOrNull { attributeName ->
+            attributeName.substringAfter(":") == "type" &&
+                    (attributeNamespaceUri(attributeName) == XML_SCHEMA_INSTANCE_NAMESPACE ||
+                            attributeName.substringBefore(":", "") == "xsi")
+        } ?: return null
+
+        val value = (attributes.getValue(attribute) as? ExactValuePattern)?.pattern?.toStringLiteral() ?: return null
+        val prefix = value.namespacePrefix()
+        val namespace = when {
+            prefix.isBlank() -> namespaceUri.orEmpty()
+            prefix == "xs" || prefix == "xsd" -> XML_SCHEMA_NAMESPACE
+            else -> attributes["xmlns:$prefix"]?.let { (it as? ExactValuePattern)?.pattern?.toStringLiteral() }.orEmpty()
+        }
+
+        return WSDLTypeName(namespace, value.localName())
+    }
+
+    internal fun wsdlTypeName(): WSDLTypeName? {
+        val namespace = wsdlTypeNamespace ?: return null
+        val name = wsdlTypeName ?: return null
+        return WSDLTypeName(namespace, name)
+    }
+
+    internal fun namespaceAttributesForXSIType(typeNamespace: String, typePrefix: String): Map<String, Pattern> {
+        val xsiNamespaceAttribute = "xmlns:xsi" to ExactValuePattern(StringValue(XML_SCHEMA_INSTANCE_NAMESPACE))
+        val typeNamespaceAttribute = when {
+            typePrefix.isBlank() -> null
+            prefixForNamespace(typeNamespace) != null -> null
+            else -> "xmlns:$typePrefix" to ExactValuePattern(StringValue(typeNamespace))
+        }
+
+        return listOfNotNull(xsiNamespaceAttribute, typeNamespaceAttribute).toMap()
+    }
+
+    internal fun prefixForNamespace(namespace: String): String? {
+        return attributes.entries.firstNotNullOfOrNull { (attributeName, pattern) ->
+            if (!attributeName.startsWith("xmlns:")) return@firstNotNullOfOrNull null
+
+            val attributeValue = (pattern as? ExactValuePattern)?.pattern?.toStringLiteral()
+            attributeName.removePrefix("xmlns:").takeIf { attributeValue == namespace }
+        }
+    }
+
+    internal fun prefixForElementNamespace(namespace: String): String? {
+        val prefix = realName.namespacePrefix()
+        return prefix.takeIf { it.isNotBlank() && namespaceUri == namespace }
+    }
+
+    internal fun availableNamespacePrefix(): String {
+        val prefixesInUse = attributes.keys
+            .filter { it.startsWith("xmlns:") }
+            .map { it.removePrefix("xmlns:") }
+            .toSet()
+
+        return (listOf("tns") + (1..100).map { index -> "ns$index" })
+            .first { prefix -> prefix !in prefixesInUse }
     }
 }
